Change setup.py to use AndroidViewClient3 namespace packaging system
Packaging is changed from previous com.dtmilano.android to simpler,
namespace contained AndroidViewClient3 (3 is added to mark that this is
a separate project and to prevent PyPI duplicate name).

When installing with `pip install AndroidViewClient3`, all of the
package files are copied to site-packages/androidviewclient3/ and
subdirectories for every subpackage (e.g. androidviewclient3/adb/).

When importing, `from AndroidViewClient3 import viewclient` is used
instead of `from com.dtmilano.android import viewclient`.